requirements.txt
bitmath==1.3.3.1
click==8.1.3
Cython==0.29.32
filterpy==1.4.5
labelme==5.1.1
numpy==1.23.3
opencv_python==4.5.5.64
pandas==1.5.0
path.py==12.5.0
pycocotools==2.0.4
pydantic==1.10.2
PyYAML==6.0
scipy==1.9.0
terminaltables==3.1.10
torchsummary==1.5.1
yacs==0.1.8
gdown
filelock
requests
uri==2.0.1
